refactor: convert logs page to typescript
