+++ src/pages/index.ts
@@ -1,6 +1,6 @@
 import { IncidentDetailsPage } from "./incident/IncidentDetails";
 import { IncidentListPage } from "./incident/IncidentListPage";
-import { LogsPage } from "./logs-page";
+import { LogsPage } from "./LogsPage";
 import { CanaryPage } from "./canary";
 import { ConfigListPage } from "./config/ConfigList";
 import { ConfigDetailsPage } from "./config/ConfigDetailsPage";
